calendar-integration/requirements.txt
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.108.0
mcp>=0.9.0
